feat: add AnnouncementBar to modern toolkit
refactored classic implementation so that both share same core functionality except
for rendering

refs coon-js/extjs-lib-comp#15
