+++ modern/src/component/AnnouncementBar.js
@@ -1,0 +1,63 @@
+/**
+ * coon.js
+ * extjs-lib-comp
+ * Copyright (C) 2022 Thorsten Suckow-Homberg https://github.com/coon-js/extjs-lib-comp
+ *
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software,
+ * and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
+ * USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+/**
+ * Modern Toolkit implementation of the AnnouncementBar.
+ */
+Ext.define("coon.comp.component.AnnouncementBar", {
+
+    extend: "coon.comp.component.AbstractAnnouncementBar",
+
+    getTemplate: function () {
+
+        return [{
+            reference: "barEl",
+            cls: "wrap",
+            tag: "div",
+            children: [{
+                cls: "msgWrap",
+                tag: "div",
+                children: [{
+                    tag: "div",
+                    cls: "msg",
+                    reference: "msgEl"
+                }, {
+                    tag: "div",
+                    cls: "yes",
+                    reference: "yesEl"
+                }, {
+                    tag: "div",
+                    cls: "no",
+                    reference: "noEl"
+                }]
+            }, {
+                tag: "div",
+                cls: "link",
+                reference: "linkEl"
+            }]
+        }];
+    }
+
+});
